Initial inclusion of compton into the Solus repositories
I hate myself, so I'm including this in the Solus repos to fox T3612.
This is often requested, but kinda poorly maintained, and upstream have
no interest in releasing new tarballs. Instead I'm using a git snapshot
for the latest commits, and will permit entry for compton based on the
2-distro alleviation rule.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>